Refactor functions and adds setup
